refactor cursor and player
