--- game_mechanics-master/README.md
@@ -1,2 +1,0 @@
-# game_mechanics
-[Demo Link](https://sackofcodetatoes.github.io/game_mechanics/)
