--- game_mechanics-master/index.html
@@ -1,14 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-  <meta charset="utf-8" />
-  <meta http-equiv="X-UA-Compatible" content="IE=edge">
-  <title>Game Mechanics</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <link rel="stylesheet" type="text/css" media="screen" href="style/main.css" />
-</head>
-<body>
-  <canvas id="game-canvas" width="1280" height="720"></canvas>
-</body>
-<script src="javascript/bundle.js"></script>
-</html>
