--- game_mechanics-master/javascript/platform.js
@@ -1,11 +1,0 @@
-import GameEntitiy from "./game_entity.js";
-
-class Platform extends GameEntitiy {
-  constructor(options){
-    super(options);
-  }
-
-
-}
-
-export default Platform;
