--- game_mechanics-master/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "game_mechanics",
-  "version": "1.0.0",
-  "description": "",
-  "main": "index.js",
-  "directories": {
-    "doc": "docs"
-  },
-  "dependencies": {
-    "webpack": "^4.20.2"
-  },
-  "devDependencies": {
-    "webpack-cli": "^3.1.2"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack --watch javascript/entry.js -o javascript/bundle.js --mode=development"
-  },
-  "author": "",
-  "license": "ISC"
-}
