--- game_mechanics-master/webpack.config.js
@@ -1,7 +1,0 @@
-module.exports = {
-  entry: "./javascript/entry.js",
-  output: {
-    filename: "./javascript/bundle.js"
-  },
-  devtool: 'source-map',
-};
