+++ javascript/camera.js
