+++ javascript/game_entity.js
@@ -2,37 +2,70 @@ class GameEntity {
   constructor(options){
     this.x = options.x;
     this.y = options.y;
-    this.context = options.context;
-    this.color = options.color;
-    this.x_len = options.x_len;
-    this.y_len = options.y_len;
-    this.draw = this.draw.bind(this);
-    this.hspd = 0;
+    this.xLen = options.xLen;
+    this.yLen = options.yLen;
     this.vspd = 0;
-    this.active = true;
-    this.faceDir = 1;
-  }
-  draw(viewPort) {
-    this.context.fillStyle = this.color;
-    this.context.fillRect(this.x - viewPort.x, this.y - viewPort.y, this.x_len, this.y_len);
+    this.hspd = 0;
+    this.physicsObj = false || options.physicsObj;
+    this.defaultColor = options.color || 'gray'
+
+    this.context = options.context;
+    this.platformCollision = options.platformCollision;
+    this.physicsCollision = options.physicsCollision;
+
+
+    this.draw = this.draw.bind(this);
+    this.stepCollisionCheck = this.stepCollisionCheck.bind(this);
   }
 
-  move() {
-    this.x += this.hspd;
-    this.y += this.vspd;
+  draw(viewPort){
+    //check if sprite, else draw green
+    this.context.fillStyle = this.defaultColor;
+    this.context.fillRect(this.x - viewPort.x, this.y - viewPort.y, this.xLen, this.yLen);
   }
-  
-  positionMeeting(x, y, otherObj){
-    //return true or false if new position intersects other objects position
 
-    //check right side
-    if((x + this.x_len > otherObj.x && x < otherObj.x + otherObj.x_len) && 
-      (y + this.y_len > otherObj.y && y < otherObj.y + otherObj.y_len)
+  update(viewPort){
+    if(this.physicsObj){
+      this.stepCollisionCheck();
+    }
+    this.draw(viewPort);
+  }
+
+  stepCollisionCheck(){
+    if (!this.platformCollision(this.x + this.hspd, this.y, this) && !this.physicsCollision(this.x + this.hspd, this.y, this)) {
+      this.x += this.hspd;
+    } else {
+      let sign = 1;
+      this.hspd < 0 ? sign = -1 : sign = sign;
+      while (!this.platformCollision(this.x + sign, this.y, this) && !this.physicsCollision(this.x + sign, this.y, this)) {
+        this.x += sign;
+      }
+    }
+
+    this.hspd = 0;
+
+    if (!this.platformCollision(this.x, this.y + this.vspd, this) && !this.physicsCollision(this.x, this.y + this.vspd, this)) {
+      this.y += this.vspd;
+    } else {
+      let sign = 1;
+      this.vspd < 0 ? sign = -1 : sign = sign;
+      while (!this.platformCollision(this.x, this.y + sign, this) && !this.physicsCollision(this.x, this.y + sign, this)) {
+        this.y += sign;
+      }
+
+
+      this.vspd = 0;
+    }
+  }
+
+  positionMeeting(x, y, obj){
+    if ((x + this.xLen > obj.x && x < obj.x + obj.xLen) &&
+      (y + this.yLen > obj.y && y < obj.y + obj.yLen)
     ) {
-      // console.log('aw shoot');
       return true;
-    }// end of if
+    } // end of if
     return false;
   }
 }
-module.exports = GameEntity;
+
+export default GameEntity;
