+++ old_javascript/coin.js
