+++ old_javascript/collision_check.js
